+++ extensions/google-maps-search/src/utils/url.ts
@@ -1,22 +1,4 @@
 /**
- * Corresponds to Google Map's four possible modes of travel.
- */
-export enum TravelMode {
-  Driving = "driving",
-  Walking = "walking",
-  Bicycling = "bicycling",
-  Transit = "transit",
-}
-
-/**
- * Corresponds to the preferences defined in package.json.
- */
-export interface Preferences {
-  homeAddress: string;
-  preferredMode: string;
-}
-
-/**
  * Given an origin, a destination, and a travel mode, returns a direction url according to the following specs:
  * https://developers.google.com/maps/documentation/urls/get-started
  *
